Improve table parsing
